feature: Build ESLint docset
This commit adds everything necessary to build the ESLint docset: It
clone necessary repositories, runs build scripts, applies some
transformations to make the documentation work in Dash and finally
creates the docset itself.

ESLint version: `8.1.0`.

There are some limitations:

* Only the user guide, developer guide and maintainer-guide are
currently included in the docset.
* Some links within the original ESLint documentation are not relative
links, but absolute links to their website. These need to be fixed in
the changed in the original documentation